+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 2.3.7
+- Add Right-Clicking on Baby Villager With Weakness Potion with a Carved Pumpkin to turn them into a Pumpkin Warden
 
 # 2.3.6
 - Fix Vine like blocks not being compostable on Forge